First_glance/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.28.0
pillow==9.5.0
joblib==1.3.2